data_processing/requirements.txt
aioconsole==0.6.2
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
bleak==0.20.2
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
executing==1.2.0
fonttools==4.42.1
ipykernel==6.25.1
ipython==8.14.0
jedi==0.19.0
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.5
matplotlib==3.7.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
numpy==1.24.3
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.74
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pygame==2.5.2
Pygments==2.16.1
pynput==1.7.6
pyobjc-core
pyobjc-framework-ApplicationServices
pyobjc-framework-Cocoa
pyobjc-framework-CoreBluetooth
pyobjc-framework-libdispatch
pyobjc-framework-Quartz
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.1
six==1.16.0
spherov2==0.12
stack-data==0.6.2
tornado==6.3.3
traitlets==5.9.0
transforms3d==0.4.1
tzdata==2023.3
wcwidth==0.2.6
websockets==11.0.3
